Remove performance and accuracy jobs, migrate test_bench_one_batch
- Split test_bench_one_batch.py into 1gpu and 2gpu versions
- Move to test/registered/perf/ with CI registration
- Remove performance-test-* jobs (tests now in stage-b suites)
- Remove accuracy-test-* jobs (tests now in stage-b suites)
- Remove from pr-test-finish needs list
